Fixed vertex index lookup table sorting
This is a cheeky and pesky bug that was fixed thanks to @edebrouwer. The
issue was that the vertex indices were treated as a lookup table, i.e. I
expected position `i` to give me the rank of vertex `i` wrt. filtration.
What I actually *used*, however, was a lookup into the *sorted* indices,
i.e. just checking which vertex *has* rank `i`. This is quite obviously
not what we want here.